Move sentinels into one file
