+++ vllm/v1/worker/gpu_worker.py
@@ -4,9 +4,7 @@
 
 import gc
 import os
-import threading
 from collections.abc import Callable
-from concurrent.futures import FIRST_EXCEPTION, ThreadPoolExecutor, wait
 from contextlib import AbstractContextManager, nullcontext
 from functools import partial
 from types import NoneType
@@ -21,12 +19,10 @@ import vllm.envs as envs
 from vllm.config import CUDAGraphMode, VllmConfig, set_current_vllm_config
 from vllm.config.compilation import CompilationMode
 from vllm.distributed import (
-    cleanup_dist_env_and_memory,
     ensure_model_parallel_initialized,
     init_distributed_environment,
     set_custom_all_reduce,
 )
-from vllm.distributed.device_communicators.cuda_communicator import CudaCommunicator
 from vllm.distributed.ec_transfer import ensure_ec_transfer_initialized
 from vllm.distributed.eplb.eplb_utils import override_envs_for_eplb
 from vllm.distributed.kv_transfer import (
@@ -36,9 +32,7 @@ from vllm.distributed.kv_transfer import (
     has_kv_transfer_group,
 )
 from vllm.distributed.parallel_state import (
-    GroupCoordinator,
     Handle,
-    get_all_model_groups,
     get_pcp_group,
     get_pp_group,
     get_tp_group,
@@ -57,13 +51,13 @@ from vllm.utils.mem_utils import MemorySnapshot, format_gib, memory_profiling
 from vllm.utils.torch_utils import set_random_seed
 from vllm.v1.core.sched.output import GrammarOutput, SchedulerOutput
 from vllm.v1.engine import ReconfigureDistributedRequest, ReconfigureRankType
-from vllm.v1.engine.base_sentinel import BaseSentinel
 from vllm.v1.kv_cache_interface import KVCacheConfig, KVCacheSpec
 from vllm.v1.outputs import (
     AsyncModelRunnerOutput,
     DraftTokenIds,
     ModelRunnerOutput,
 )
+from vllm.v1.sentinel import WorkerSentinel
 from vllm.v1.utils import compute_iteration_details, report_usage_stats
 from vllm.v1.worker.utils import is_residual_scattered_for_sp
 from vllm.v1.worker.worker_base import WorkerBase
@@ -108,126 +102,6 @@ class AsyncIntermediateTensors(IntermediateTensors):
         if name == "tensors" and not object.__getattribute__(self, "_comm_waited"):
             object.__getattribute__(self, "wait_for_comm")()
         return object.__getattribute__(self, name)
-
-
-class WorkerSentinel(BaseSentinel):
-    def __init__(
-        self,
-        vllm_config: VllmConfig,
-        pause_event: threading.Event,
-        init_distributed_env_callback: Callable,
-        clear_input_batch_callback: Callable,
-        device: torch.cuda.device,
-    ):
-        self.dp_rank = vllm_config.parallel_config.data_parallel_rank
-        self.tp_rank = get_tp_group().rank_in_group
-        self.pp_rank = get_pp_group().rank_in_group
-        identity = f"PP{self.pp_rank}_TP{self.tp_rank}"
-        super().__init__(
-            upstream_cmd_addr=vllm_config.fault_tolerance_config.worker_cmd_addr,
-            downstream_cmd_addr=None,
-            sentinel_identity=identity.encode(),
-            sentinel_tag=f"{self.dp_rank}_{identity}",
-            vllm_config=vllm_config,
-        )
-        self.init_distributed_env_callback = init_distributed_env_callback
-        self.clear_input_batch_callback = clear_input_batch_callback
-        self.device = device
-
-        self.pause_event = pause_event
-        self.communicator_aborted = False
-        torch.cuda.set_device(self.device)
-        threading.Thread(
-            target=self.run, daemon=True, name="WorkerSentinelMonitorThread"
-        ).start()
-
-    def run(self):
-        # Wait for fault tolerance instructions from EngineCoreSentinel
-        while not self.sentinel_dead:
-            self.poll_and_execute_upstream_cmd()
-
-    def pause(self, timeout: int = 1, **kwargs) -> bool:
-        soft_pause = kwargs.get("soft_pause", False)
-        if soft_pause:
-            self._set_device_communicator_status(False)
-            self.pause_event.set()
-            self.logger("Pause signal sent.")
-            return True
-        # Abort all NCCL communicators and
-        # process groups in parallel using a thread pool.
-        if self.communicator_aborted:
-            return True
-        self.pause_event.set()
-        self._set_device_communicator_status(False)
-        torch.cuda.set_device(self.device)
-        model_groups = get_all_model_groups()
-        futures = []
-
-        def _abort_nccl_comm(group: GroupCoordinator):
-            if group.device_communicator is not None:
-                device_comm = cast(CudaCommunicator, group.device_communicator)
-                nccl_comm = device_comm.pynccl_comm
-                assert nccl_comm is not None
-                nccl_comm.nccl_abort_comm()
-
-        def _abort_process_group(group: GroupCoordinator):
-            backend = group.device_group._get_backend(self.device)
-            backend.abort()
-
-        executor = ThreadPoolExecutor(max_workers=len(model_groups) * 2)
-        try:
-            for group in model_groups:
-                futures.append(executor.submit(_abort_nccl_comm, group))
-                futures.append(executor.submit(_abort_process_group, group))
-
-            done, not_done = wait(futures, timeout=timeout, return_when=FIRST_EXCEPTION)
-            if not_done:
-                self.logger(
-                    "%d abort calls did not finish in total %s seconds",
-                    len(not_done),
-                    timeout,
-                    level="warning",
-                )
-        finally:
-            executor.shutdown(wait=False, cancel_futures=True)
-
-        exception_count = sum(1 for f in done if f.exception() is not None)
-        self.communicator_aborted = len(not_done) == 0 and exception_count == 0
-        if self.communicator_aborted:
-            cleanup_dist_env_and_memory()
-            self.logger("Communicators are aborted.")
-        else:
-            self.logger(
-                "Communicator abort failed: %d NCCL comm abort calls timed out,"
-                " %d tasks threw exceptions. This may leave NCCL communicators "
-                "or process groups in an inconsistent state. Subsequent "
-                "distributed operations could be unsafe.",
-                len(not_done),
-                exception_count,
-                level="error",
-            )
-        return self.communicator_aborted
-
-    def _set_device_communicator_status(self, active: bool):
-        model_groups = get_all_model_groups()
-        for group in model_groups:
-            if group.device_communicator is not None:
-                device_comm = cast(CudaCommunicator, group.device_communicator)
-                nccl_comm = device_comm.pynccl_comm
-                assert nccl_comm is not None
-                nccl_comm.available = active
-                nccl_comm.disabled = not active
-
-    def retry(self, timeout: int = 1, **kwargs) -> bool:
-        if self.communicator_aborted:
-            torch.cuda.set_device(self.device)
-            with set_current_vllm_config(self.vllm_config):
-                self.init_distributed_env_callback()
-                self.communicator_aborted = False
-            torch.cuda.synchronize()
-        self.clear_input_batch_callback()
-        self.pause_event.clear()
-        return True
 
 
 class Worker(WorkerBase):
